tidying up and adding gwas specific tools

+++ harmonisation/copy_files_to_ftp.sh
@@ -18,38 +18,52 @@ harmonised_dir=$4
 readme=$5
 
 ss_path=$(find $ftp_path -type d -name *_$study)
-formatted_file_path=$(find $formatted_dir -type f -name *$study*)
-harmonised_file_path=$(find $harmonised_dir -type f -name *$study*)
- 
-formatted_file=$(basename $formatted_file_path)
-harmonised_file=$(basename $harmonised_file_path)
+formatted_file_path=$(find $formatted_dir -type f -name "*$study*")
+harmonised_file_path=$(find $harmonised_dir -type f -name "*$study*")
+
+formatted_file=$(basename $formatted_file_path .tsv)
+harmonised_file=$(basename $harmonised_file_path .tsv)
 
 
 # If no path with study accession:
 
 if [ ! -d "$ss_path" ]; then
 	echo "No directory for $study"
-	echo "Checking database to try and generate..."
-        ss_dir=$(python2 generate_sumstats_dir_name.py --study $study)
-	echo $ss_dir
+#	echo "Checking database to try and generate..."
+#        ss_dir=$(python2 /nfs/spot/data/test/sumstats/formatting_tools/generate_sumstats_dir_name.py --study $study)
+#	echo $ss_dir
 
-	if [[ $ss_dir =~ "ERROR" ]]; then
-		echo "ERROR: Could not find ${study} in the database" 
-	else
-		echo "making dir: ${ftp_path}${ss_dir}"
-		mkdir $ftp_path$ss_dir
-		ss_path=$ftp_path$ss_dir
-	fi
+#	if [[ $ss_dir =~ "ERROR" ]]; then
+#		echo "ERROR: Could not find ${study} in the database" | mail -s "File failed to copy" jhayhurst@ebi.ac.uk
+#	else
+#		echo "making dir: ${ftp_path}${ss_dir}"
+#		mkdir $ftp_path$ss_dir
+#		ss_path=$ftp_path$ss_dir
+#	fi
 else
 	# If path with study accession, get the name of that directory:
 	ss_dirname=$(echo $ss_path | rev | cut -f1 -d "/" | rev)
-	echo "$ss_dirname already exists" 
+	echo "$ss_dirname exists" 
 fi
 
 if [ -d "$ss_path" ]; then
-	mkdir $ss_path/harmonised
-	echo "copying files: $formatted_file_path and $harmonised_file_path to $ss_path"
-	gzip < $formatted_file_path > $ss_path/harmonised/$formatted_file.f.gz
-	gzip < $harmonised_file_path > $ss_path/harmonised/$harmonised_file.h.gz 
-	cp $readme $ss_path/harmonised/
+	# if harmonised files don't exist:
+	mkdir -p $ss_path/harmonised
+        #rm -vf $ss_path/harmonised/*.tsv.f.gz
+        #rm -vf $ss_path/harmonised/*.tsv.h.gz
+        if [ ! -e "${ss_path}/harmonised/${formatted_file}.f.tsv.gz" ]; then
+		gzip < $formatted_file_path > $ss_path/harmonised/$formatted_file.f.tsv.gz
+		echo "copying: $formatted_file_path to $ss_path/harmonised/"
+	else
+		echo "$formatted_file_path already exists"
+	fi
+        if [ ! -e "$ss_path/harmonised/${harmonised_file}.h.tsv.gz" ]; then
+                gzip < $harmonised_file_path > $ss_path/harmonised/$harmonised_file.h.tsv.gz
+                echo "copying: $harmonised_file_path to $ss_path/harmonised/"
+        else
+                echo "$harmonised_file_path already exists"
+        fi
+	if [ ! -e "${ss_path}/harmonised/readme.txt" ]; then
+		cp $readme $ss_path/harmonised/
+	fi
 fi
